refactor: Category, WishList pages, seperate some constants, util functions, component
