+++ client/src/components/index.js
@@ -6,3 +6,5 @@ export { default as FormZoneCodeInput } from './Sign/FormZoneCodeInput';
 export { default as NavigationBar } from './NavigationBar';
 export { default as Footer } from './Footer';
 export { default as SwitchingpageScrollToTop } from './SwitchigPageScrollToTop';
+export { default as NoProduct } from './NoProduct';
+export { default as SizeButton } from './SizeButton';
