+++ client/src/constants/brands.js
@@ -1,0 +1,19 @@
+const BRANDS = [
+  { en: 'nike', kr: '나이키' },
+  { en: 'adidas', kr: '아디다스' },
+  { en: 'newBalance', kr: '뉴발란스' },
+  { en: 'sketchers', kr: '스케쳐스' },
+  { en: 'ecco', kr: '에코' },
+  { en: 'crocs', kr: '크록스' },
+  { en: 'asics', kr: '아식스' },
+  { en: 'descente', kr: '데상트' },
+  { en: 'converse', kr: '컨버스' },
+  { en: 'kumkang', kr: '금강제화' },
+  { en: 'babara', kr: '바바라' },
+  { en: 'saera', kr: '새라' },
+  { en: 'tandy', kr: '텐디' },
+  { en: 'rockport', kr: '락포트' },
+  { en: 'soda', kr: '소다' },
+];
+
+export default BRANDS;
