+++ client/src/constants/categories.js
@@ -1,0 +1,9 @@
+const CATEGORIES = {
+  sneakers: '운동화',
+  sandal: '샌달',
+  slipper: '슬리퍼',
+  walking: '워킹화',
+  shoes: '구두',
+};
+
+export default CATEGORIES;
