+++ client/src/constants/colors.js
@@ -1,0 +1,17 @@
+const COLORS = [
+  { color: '#8D429F', en: 'purple', kr: '보라색' },
+  { color: '#000', en: 'black', kr: '검정색' },
+  { color: '#E7352B', en: 'red', kr: '빨간색' },
+  { color: '#F36B26', en: 'orange', kr: '주황색' },
+  { color: '#1790C8', en: 'blue', kr: '파란색' },
+  { color: '#ffffff', en: 'white', kr: '흰색' },
+  { color: '#825D41', en: 'brown', kr: '갈색' },
+  { color: '#7BBA3C', en: 'green', kr: '초록색' },
+  { color: '#FED533', en: 'yellow', kr: '노란색' },
+  { color: 'navy', en: 'navy', kr: '남색' },
+  { color: 'beige', en: 'beige', kr: '베이지' },
+  { color: '#808080', en: 'gray', kr: '회색' },
+  { color: '#F0728F', en: 'pink', kr: '분홍색' },
+];
+
+export default COLORS;
