+++ client/src/constants/gender.js
@@ -1,0 +1,6 @@
+const GENDER = [
+  { en: 'male', kr: '남성' },
+  { en: 'female', kr: '여성' },
+];
+
+export default GENDER;
