+++ client/src/constants/prices.js
@@ -1,0 +1,8 @@
+const PRICES = [
+  { rangeIdx: 0, text: '0 - 50,000원' },
+  { rangeIdx: 1, text: '50,000 - 100,000원' },
+  { rangeIdx: 2, text: '100,000 - 150,000원' },
+  { rangeIdx: 3, text: '150,000원 이상' },
+];
+
+export default PRICES;
