+++ client/src/constants/sizes.js
@@ -1,0 +1,3 @@
+const SIZES = [230, 235, 240, 245, 250, 255, 260, 265, 270, 275, 280, 285];
+
+export default SIZES;
